+++ R/from.igraph.R
@@ -1,9 +1,8 @@
-#' Transform Network from \code{igraph}  to \code{fastnet} Format
+#' Transform an \code{igraph} object to a \code{fastnet} object
 #'
-#' @description Transform \code{igraph} object to an ego-centric list form used in \code{fastnet}.
+#' @description Transform an \code{igraph} object to an ego-centric list form used in \code{fastnet}.
 #' @param net.igraph The input \code{igraph} object.
 #' @return A list containing the nodes of the network and their respective neighbors.
-#' @details Masked from \code{igraph::as_adj_list}.
 #' @author Xu Dong.
 #' @examples \dontrun{
 #' library("igraph")
@@ -12,11 +11,10 @@
 #' @export
 
 from.igraph <- function(net.igraph){
-
-  net <- igraph::as_adj_list(net.igraph)
-
+  n <- igraph::vcount(net.igraph)
+  g_0 <- igraph::set.vertex.attribute(net.igraph, "name", value=seq(n))
+  net <- unname( lapply( igraph::as_adj_list(g_0), as.numeric ) )
   net
-
 }