+++ NEWS.md
@@ -1,5 +1,16 @@
 # News
 
+## 0.1.4
+* Fix the consistancy issue when presenting a fastnet network object by removing the ‘igraph.vs’ class from the output for the function "from.adjacency" and "from.igraph"
+* Elaborate the example and the statements of the “parameters” and “details” in the "from.edgelist" documentation
+* Make the argument “ncores” consistent with other functions for "net.rewired.caveman"
+* Add a new function "to.edgelist" to Coerce a fastnet object to an edgelist representation
+* Add a new function "to.igraph" to Coerce a fastnet object to an igraph object
+* Add a new function "to.tidygraph" to Coerce a fastnet object to a tidygraph object
+* Add a new function "from.csc" to transform a sparse .csc adjacency matrix to an ego-centric list form used in \code{fastnet}
+* Add a new function "from.csr" to transform a sparse .csr adjacency matrix to an ego-centric list form used in \code{fastnet}
+* Add a new function "from.statnet" to transform a \code{network} object  to an ego-centric list form used in \code{fastnet}
+
 ## 0.1.3
 * Changes in plot labels: draw.degdist.cum.log, draw.degdist.cum, histo.dist
 * Minor changes in options and display: draw.degdist.log, draw.degdist